Finished OOP
combined all classes under one module. Broke the program into relevent classes. Removed the excess modules/py files.
